requirements.txt
asgiref==3.3.1
Django==3.1.5
django-environ==0.4.5
django-phone-field==1.8.1
djangorestframework==3.12.2
drf-extensions==0.6.0
phonenumbers==8.12.15
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
